done graphql test

+++ src/test/java/com/spring5/mongodb/CustomerRepositoryTests.java
@@ -6,26 +6,34 @@ package com.spring5.mongodb;
 
 import java.util.List;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import org.junit.Test;
 
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.springframework.core.env.Environment;
 import org.springframework.data.domain.Example;
 import org.testcontainers.containers.MongoDBContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
-//@RunWith(SpringJUnit4ClassRunner.class)
 @SpringBootTest
 @Testcontainers
 public class CustomerRepositoryTests {
+    
+    
+    @Autowired
+    Environment env;
+    
+    @Value("${app.mongodb.enabled}")
+    protected boolean mongoDbEnabled;
 
     @Container
     @ServiceConnection
-    static MongoDBContainer container = new MongoDBContainer("mongo:7.0.2");
+    static MongoDBContainer container = null;
 
     @Autowired
     CustomerRepository repository;
@@ -34,20 +42,27 @@ public class CustomerRepositoryTests {
 
     @BeforeEach
     public void setUp() {
+        if (mongoDbEnabled) {
+            container = new MongoDBContainer("mongo:7.0.2");
+            
+            repository.deleteAll();
 
-        repository.deleteAll();
-
-        dave = repository.save(new Customer("Dave", "Matthews"));
-        oliver = repository.save(new Customer("Oliver August", "Matthews"));
-        carter = repository.save(new Customer("Carter", "Beauford"));
+            dave = repository.save(new Customer("Dave", "Matthews"));
+            oliver = repository.save(new Customer("Oliver August", "Matthews"));
+            carter = repository.save(new Customer("Carter", "Beauford"));
+        }
     }
 
-    //@Test
+    @Test
     public void setsIdOnSave() {
-
-        Customer dave = repository.save(new Customer("Dave", "Matthews"));
-
-        assertThat(dave.id).isNotNull();
+        String id = null;
+        if (mongoDbEnabled) {
+            Customer dave = repository.save(new Customer("Dave", "Matthews"));
+            id= dave.id;
+            assertThat(id).isNotNull();
+        } else {
+            assertThat(id).isNull();
+        }
     }
 
     //@Test
